LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/registered-var-to-registered-animating-expected.txt


FAIL Animated registered custom properties can var() reference other animated registered custom properties across separate Animations. assert_equals: target at 25%, to at 25% expected "131.25" but got "75"
